milestone 3
This should cover all the requirements for Milestone 3. I have begun making changes to further improve the user experience and make it a more well-rounded application. Registration data now saves correctly, and adding a book saves it to both the admin side and the user side. Orientation will need additional work, as some pages open to the main screen after signing in; however, the security measures are still in place, ensuring the application remains secure.